#7572 A52 et A53 - WS Alimentation manuelle du compteur (SIRH) #7565 A52 et A53 - WS Liste des compteurs (SIRH) #7562 A52 et A53 - WS MAJ manuelle des compteurs

+++ src/main/java/nc/noumea/mairie/abs/domain/AgentAsaA52Count.java
@@ -21,7 +21,7 @@ public class AgentAsaA52Count extends AgentCount {
 
 	@NotNull
 	@Column(name = "TOTAL_MINUTES", columnDefinition = "numeric")
-	private Double totalMinutes;
+	private int totalMinutes;
 
 	@Column(name = "DATE_DEBUT")
 	@Temporal(TemporalType.TIMESTAMP)
@@ -35,11 +35,11 @@ public class AgentAsaA52Count extends AgentCount {
 	@JoinColumn(name = "ID_ORGANISATION_SYNDICALE")
 	private OrganisationSyndicale organisationSyndicale;
 
-	public Double getTotalMinutes() {
+	public int getTotalMinutes() {
 		return totalMinutes;
 	}
 
-	public void setTotalMinutes(Double totalMinutes) {
+	public void setTotalMinutes(int totalMinutes) {
 		this.totalMinutes = totalMinutes;
 	}
